move TimeParser to time.rb about symmetricity with TimeFormatter
